cleanup json schema
