+++ OpenAI-DotNet/Assistants/CreateAssistantRequest.cs
@@ -322,7 +322,7 @@ namespace OpenAI.Assistants
         /// which indicates the generation exceeded max_tokens or the conversation exceeded the max context length.
         /// </remarks>
         [JsonPropertyName("response_format")]
-        [JsonConverter(typeof(TextResponseFormatConverter))]
+        [JsonConverter(typeof(TextResponseFormatConfigurationConverter))]
         [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
         public TextResponseFormatConfiguration ResponseFormatObject { get; internal set; }
